收银台 增加 店铺促销 排除指定商品功能，修改店铺促销接口

+++ app/src/main/java/com/easygo/cashier/widget/ChooseCouponsDialog.java
@@ -86,7 +86,7 @@ public class ChooseCouponsDialog extends BaseDialog {
             @Override
             protected void convert(BaseViewHolder helper, CouponResponse item) {
                 String condition_value = item.getCondition_value();
-                int offer_value = item.getOffer_value();
+                float offer_value = item.getOffer_value();
                 String effected_at = sdf.format(item.getEffected_at());
                 String expired_at = sdf.format(item.getExpired_at());
